[NEW] metricas de diversidade calculadas para tiselac

+++ download_data.sh
@@ -3,14 +3,18 @@
 mkdir data
 
 # download LSST
-mkdir data/LSST/
-wget -P data/LSST/ http://timeseriesclassification.com/Downloads/LSST.zip
-unzip data/LSST/LSST.zip -d data/LSST/
-rm data/LSST/*.txt data/LSST/*.JPG data/LSST/*.ts
+if [ ! -d "data/LSST/" ]; then
+    mkdir data/LSST/
+    wget -P data/LSST/ http://timeseriesclassification.com/Downloads/LSST.zip
+    unzip data/LSST/LSST.zip -d data/LSST/
+    rm data/LSST/*.txt data/LSST/*.JPG data/LSST/*.ts
+fi
 
 
 # download Tiselac
-mkdir data/Tiselac/
-wget -P data/Tiselac http://timeseriesclassification.com/Downloads/Tiselac.zip
-unzip data/Tiselac/Tiselac.zip -d data/Tiselac/
-rm data/Tiselac/*.zip data/Tiselac/*.png
+if [ ! -d "data/Tiselac" ]; then
+    mkdir data/Tiselac/
+    wget -P data/Tiselac http://timeseriesclassification.com/Downloads/Tiselac.zip
+    unzip data/Tiselac/Tiselac.zip -d data/Tiselac/
+    rm data/Tiselac/*.zip data/Tiselac/*.png
+fi
